+++ DoofesZeug.Library/Src/Attributes/Validation/MinAttribute.cs
@@ -13,13 +13,13 @@ namespace DoofesZeug.Attributes.Validation
         /// <value>
         /// The minimum.
         /// </value>
-        public int Min { get; private set; }
+        public double Min { get; private set; }
 
 
         /// <summary>
         /// Initializes a new instance of the <see cref="MinAttribute"/> class.
         /// </summary>
         /// <param name="min">The minimum.</param>
-        public MinAttribute( int min ) => this.Min = min;
+        public MinAttribute( double min ) => this.Min = min;
     }
 }
